requirements.txt
django<3.3
whitenoise
watchdog[watchmedo]
jinja2
django-environ
django-ninja
django-allauth
django_use_email_as_username
waitress
psycopg2-binary
pyyaml
pypugjs
markdown-it-py
